requirements.txt
crewai==1.9.3
langchain-google-genai==4.2.0
python-dotenv==1.1.1
requests==2.32.5
streamlit==1.54.0
